+++ src/lib/__init__.py
@@ -1,8 +1,8 @@
-from .touchstone import Touchstone, extrapolate_sparams_to_dc, get_sparam_name, sparam_to_timedomain
+from .sparam_helpers import extrapolate_sparams_to_dc, get_sparam_name, sparam_to_timedomain
 from .si import Si, SiFmt
-from .structs import LoadedSParamFile, PlotData, PlotDataQuantity
+from .structs import SParamFile, PlotData, PlotDataQuantity
 from .plot import PlotHelper
-from .appsettings import AppSettings
+from .appsettings import app_settings
 from .utils import get_unique_short_filename, is_windows
 from .expressions import ExpressionParser
 from .tkinter_helpers import TkText, TkCommon